rustc_hir_typeck: fix paths and partially mv files
